refactor: address code review enhancements
- Move path utilities from constants to utils package
- Remove ModelDirectory global variable
- Simplify baseDir configuration in values.yaml
- Update documentation and test expectations
