+++ ai-services/assets/catalog/podman/values.yaml
@@ -7,7 +7,7 @@ backend:
   image: icr.io/ai-services-cicd/ai-services:0.0.4
   runtime: ""
   adminPasswordHash: ""
-  baseDir: "/var/lib/ai-services"
+  baseDir: ""
   podman:
     uri: "/run/podman/podman.sock"
 
@@ -17,7 +17,6 @@ db:
   user: "admin"
   database: "ai_services"
   password: ""
-  dataPath: ""  # Will be set dynamically based on backend.baseDir
   resources:
     requests:
       memory: "512Mi"
