+++ ai-services/cmd/ai-services/cmd/application/model/download.go
@@ -6,9 +6,12 @@ import (
 	"github.com/project-ai-services/ai-services/internal/pkg/cli/helpers"
 	"github.com/project-ai-services/ai-services/internal/pkg/logger"
 	"github.com/project-ai-services/ai-services/internal/pkg/runtime/types"
+	"github.com/project-ai-services/ai-services/internal/pkg/utils"
 	"github.com/project-ai-services/ai-services/internal/pkg/vars"
 	"github.com/spf13/cobra"
 )
+
+var modelDirectory string
 
 var downloadCmd = &cobra.Command{
 	Use:   "download",
@@ -29,7 +32,7 @@ func init() {
 	_ = downloadCmd.MarkFlagRequired("template")
 	downloadCmd.Flags().StringVar(&vars.ToolImage, "tool-image", vars.ToolImage, "Tool container image used for downloading the model (for development purposes only)")
 	_ = downloadCmd.Flags().MarkHidden("tool-image")
-	downloadCmd.Flags().StringVar(&vars.ModelDirectory, "dir", vars.ModelDirectory, "Directory to download the model files")
+	downloadCmd.Flags().StringVar(&modelDirectory, "dir", utils.GetModelsPath(), "Directory to download the model files")
 }
 
 func download(cmd *cobra.Command) error {
@@ -46,7 +49,7 @@ func download(cmd *cobra.Command) error {
 	}
 	logger.Infoln("Downloaded Models in application template" + templateName + ":")
 	for _, model := range models {
-		err := helpers.DownloadModel(model, vars.ModelDirectory)
+		err := helpers.DownloadModel(model, modelDirectory)
 		if err != nil {
 			return fmt.Errorf("failed to download model: %w", err)
 		}
