+++ ai-services/internal/pkg/vars/var.go
@@ -4,7 +4,6 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/project-ai-services/ai-services/internal/pkg/constants"
 	"github.com/project-ai-services/ai-services/internal/pkg/runtime"
 )
 
@@ -17,8 +16,6 @@ var (
 	// SpyreCardAnnotationRegex -> ai-services.io/<containerName>--spyre-cards.
 	SpyreCardAnnotationRegex = regexp.MustCompile(`^ai-services\.io\/([A-Za-z0-9][-A-Za-z0-9_.]*)--spyre-cards$`)
 	ToolImage                = "icr.io/ai-services/tools:0.7"
-	// ModelDirectory is the directory for AI models, uses GetModelsPath() for dynamic resolution.
-	ModelDirectory = constants.GetModelsPath()
 )
 
 type Label string
